Fix Batches 1-10: TicketSubsystem sanitize bugs, IntervalMinutes braces, freeze/unfreeze UID, mute int32 overflow, template substitution order, WebSocket RFC violations, totalPages overflow, BanDatabase FDateTime sentinel, appeal TOCTOU, BanBridgeConfig variable order
Agent-Logs-Url: https://github.com/Yamahasxviper/DiscordBridge-/sessions/05c93e30-6d83-4a16-82d7-c7bfb47c93a8

+++ Mods/BanSystem/Source/BanSystem/Private/BanRestApi.cpp
@@ -1324,7 +1324,7 @@ void UBanRestApi::RegisterRoutes()
             Root->SetNumberField(TEXT("total"),      Total);
             Root->SetNumberField(TEXT("page"),       Page);
             Root->SetNumberField(TEXT("pageSize"),   Limit);
-            Root->SetNumberField(TEXT("totalPages"), (Total + Limit - 1) / FMath::Max(Limit, 1));
+            Root->SetNumberField(TEXT("totalPages"), (static_cast<int64>(Total) + Limit - 1) / FMath::Max(Limit, 1));
             Root->SetArrayField (TEXT("entries"),    Arr);
             Done(BanJson::Json(BanJson::ObjectToString(Root)));
             return true;
